Move server/etcd_quota_provider.go into quota/etcd package
Makes it optional to import.
